+++ packages/firebase/src/lib/initializeFirebase.ts
@@ -1,15 +1,20 @@
 import { ApiConfig } from "@dzangolab/fastify-config";
-import { initializeApp, credential } from "firebase-admin";
+import { FastifyInstance } from "fastify";
+import { initializeApp, credential, apps } from "firebase-admin";
 
-const initializeFirebase = (config: ApiConfig) => {
-  if (config.firebase && config.firebase.projectId) {
-    initializeApp({
-      credential: credential.cert({
-        projectId: config.firebase.projectId,
-        privateKey: config.firebase.privateKey,
-        clientEmail: config.firebase.clientEmail,
-      }),
-    });
+const initializeFirebase = (config: ApiConfig, fastify: FastifyInstance) => {
+  if (apps.length === 0) {
+    try {
+      initializeApp({
+        credential: credential.cert({
+          projectId: config.firebase.credentials.projectId,
+          privateKey: config.firebase.credentials.privateKey,
+          clientEmail: config.firebase.credentials.clientEmail,
+        }),
+      });
+    } catch {
+      fastify.log.error("Failed to initialize firebase");
+    }
   }
 };